requirements.txt
PyYAML>=6.0.3
pytest==8.4.2
